chore(tiles-gc): split into src/ and add tsconfig references
Move gc.ts and gc.test.ts under src/, update cli.ts import path, add
tsconfig project references so pnpm typecheck covers tiles-gc, and add
@world-history-map/tiles workspace dependency.
