+++ scripts/tiles-gc/package.json
@@ -8,6 +8,9 @@
     "test": "vitest run",
     "test:coverage": "vitest run --coverage"
   },
+  "dependencies": {
+    "@world-history-map/tiles": "workspace:*"
+  },
   "devDependencies": {
     "@types/node": "^25.5.0",
     "@vitest/coverage-v8": "^4.1.2",
