+++ scripts/tiles-gc/vitest.config.ts
@@ -1,9 +1,20 @@
+import path from 'node:path';
 import { defineConfig } from 'vitest/config';
 
 export default defineConfig({
+  resolve: {
+    alias: {
+      '@': path.resolve(import.meta.dirname, './src'),
+    },
+  },
   test: {
     globals: true,
     environment: 'node',
-    include: ['*.test.ts'],
+    include: ['src/**/*.test.ts'],
+    coverage: {
+      provider: 'v8',
+      include: ['src/**/*.ts'],
+      exclude: ['src/**/*.d.ts', 'src/**/*.{test,spec}.ts'],
+    },
   },
 });
